requirements.txt
# --- Core Frameworks ---
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-google-genai>=1.0.4

# --- Embeddings & Models ---
langchain-huggingface>=0.1.0
sentence-transformers==2.7.0
huggingface-hub==0.23.0

# --- Vector Store & Retrievers ---
langchain-chroma>=0.1.1
chromadb==0.5.0
rank-bm25==0.2.2

# --- Ollama Support ---
langchain-ollama>=0.1.0


fastapi
uvicorn[standard]
python-dotenv

langchain_qdrant
qdrant-client
sentence-transformers
pydantic